passport.controller and auth.router 2 Typescript
