+++ server/package.json
@@ -33,6 +33,7 @@
     "@types/express-session": "^1.17.4",
     "@types/node": "^15.12.1",
     "@types/passport": "^1.0.6",
+    "@types/passport-google-oauth20": "^2.0.9",
     "@types/uuid": "^8.3.1",
     "@typescript-eslint/eslint-plugin": "^4.28.0",
     "@typescript-eslint/parser": "^4.28.0",
